♻️ Extract the base test suite as single module.
